requirements.txt
flask==2.3.2
telethon==1.28.5
python-dotenv==1.0.0 